@@ -1,110 +1,117 @@
 {
   "name": "@uzh-bf/design-system",
-  "version": "1.6.1",
+  "version": "3.0.0-alpha.35",
   "main": "dist/index.js",
-  "module": "dist/index.esm.js",
-  "typings": "dist/index.d.ts",
   "files": [
     "dist",
     "src"
   ],
-  "scripts": {
-    "analyze": "size-limit --why",
-    "build": "npm-run-all build:dts build:copy",
-    "build:dts": "dts build --entry src/index.ts --entry src/constants.js",
-    "build:copy": "node src/copy.js",
-    "lint": "dts lint",
-    "size": "size-limit",
-    "dev": "ladle serve",
-    "deploy": "ladle build",
-    "preview": "ladle build && ladle preview -p 61000",
-    "format": "prettier --write src/",
-    "format:check": "prettier --check src/",
-    "test": "dts test --passWithNoTests",
-    "prerelease": "npm run build",
-    "release": "standard-version",
-    "release:dry": "standard-version --dry-run",
-    "check": "tsc --noEmit"
-  },
-  "jest": {
-    "testEnvironment": "jsdom"
-  },
-  "engines": {
-    "node": ">=16"
-  },
-  "size-limit": [
-    {
-      "path": "dist/index.cjs.production.min.js",
-      "limit": "10 KB"
-    },
-    {
-      "path": "dist/index.esm.js",
-      "limit": "10 KB"
-    }
-  ],
-  "devDependencies": {
-    "@ladle/react": "2.15.0",
-    "@rollup/plugin-alias": "5.0.1",
-    "@rollup/plugin-html": "1.0.3",
-    "@size-limit/preset-small-lib": "7.0.8",
-    "@tailwindcss/aspect-ratio": "0.4.2",
-    "@tailwindcss/forms": "0.5.3",
-    "@tailwindcss/typography": "0.5.9",
-    "@tsconfig/create-react-app": "1.0.3",
-    "@tsconfig/recommended": "1.0.1",
-    "@types/node": "^18.11.9",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "@vitejs/plugin-react-swc": "3.3.2",
-    "autoprefixer": "10.4.16",
-    "dts-cli": "1.6.3",
-    "eslint": "8.44.0",
-    "husky": "8.0.3",
-    "npm-run-all": "4.1.5",
-    "prettier": "2.8.1",
-    "prettier-plugin-organize-imports": "3.2.2",
-    "prettier-plugin-tailwindcss": "0.3.0",
-    "rollup-plugin-postcss": "4.0.2",
-    "size-limit": "7.0.8",
-    "standard-version": "9.5.0",
-    "ts-jest": "29.0.5",
-    "tslib": "2.6.0",
-    "typescript": "5.1.6"
-  },
-  "peerDependencies": {
-    "@fortawesome/fontawesome-svg-core": "^6.3.0",
-    "@fortawesome/free-regular-svg-icons": "^6.3.0",
-    "@fortawesome/free-solid-svg-icons": "^6.3.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "autoprefixer": "^10.0.0",
-    "dayjs": "^1.0.0",
-    "formik": "^2.2.9",
-    "postcss-import": "^15.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tailwind-merge": "^1.8.1",
-    "tailwindcss": "^3.3.0",
-    "tailwindcss-radix": "^2.8.0",
-    "yup": "^1.0.0"
-  },
   "dependencies": {
-    "@radix-ui/react-checkbox": "1.0.4",
+    "@hookform/resolvers": "3.9.0",
+    "@radix-ui/react-accordion": "1.2.0",
+    "@radix-ui/react-alert-dialog": "1.1.1",
+    "@radix-ui/react-aspect-ratio": "1.1.0",
+    "@radix-ui/react-avatar": "1.1.0",
+    "@radix-ui/react-checkbox": "1.1.1",
     "@radix-ui/react-collapsible": "1.0.3",
+    "@radix-ui/react-context-menu": "2.2.1",
     "@radix-ui/react-dialog": "1.0.4",
     "@radix-ui/react-dropdown-menu": "2.0.5",
-    "@radix-ui/react-label": "2.0.2",
-    "@radix-ui/react-navigation-menu": "1.1.3",
+    "@radix-ui/react-hover-card": "1.1.1",
+    "@radix-ui/react-label": "2.1.0",
+    "@radix-ui/react-menubar": "1.1.1",
+    "@radix-ui/react-navigation-menu": "1.2.0",
+    "@radix-ui/react-popover": "1.1.1",
     "@radix-ui/react-progress": "1.0.3",
+    "@radix-ui/react-radio-group": "1.2.0",
+    "@radix-ui/react-scroll-area": "1.1.0",
     "@radix-ui/react-select": "1.2.2",
+    "@radix-ui/react-separator": "1.1.0",
     "@radix-ui/react-slider": "1.1.2",
+    "@radix-ui/react-slot": "1.1.0",
     "@radix-ui/react-switch": "1.0.3",
     "@radix-ui/react-tabs": "1.0.4",
     "@radix-ui/react-toast": "1.1.4",
+    "@radix-ui/react-toggle": "1.1.0",
+    "@radix-ui/react-toggle-group": "1.1.0",
     "@radix-ui/react-tooltip": "1.0.6",
+    "cmdk": "1.0.0",
+    "date-fns": "3.6.0",
+    "embla-carousel-react": "8.1.8",
+    "input-otp": "1.2.4",
+    "next-themes": "0.3.0",
     "react-colorful": "5.6.1",
-    "react-countdown": "2.3.5"
+    "react-countdown": "2.3.5",
+    "react-day-picker": "8.10.1",
+    "react-hook-form": "7.52.2",
+    "react-resizable-panels": "2.0.22",
+    "recharts": "2.12.7",
+    "sonner": "1.5.0",
+    "vaul": "0.9.1",
+    "zod": "3.23.8"
+  },
+  "devDependencies": {
+    "@ladle/react": "4.1.0",
+    "@tailwindcss/aspect-ratio": "0.4.2",
+    "@tailwindcss/forms": "0.5.7",
+    "@tailwindcss/typography": "0.5.13",
+    "@types/node": "^20.14.14",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "7.15.0",
+    "@typescript-eslint/parser": "7.15.0",
+    "@vitejs/plugin-react-swc": "3.7.0",
+    "autoprefixer": "10.4.20",
+    "eslint": "8.57.0",
+    "eslint-plugin-react-hooks": "4.6.2",
+    "eslint-plugin-react-refresh": "0.4.9",
+    "npm-run-all": "4.1.5",
+    "postcss": "8.4.41",
+    "prettier": "3.3.3",
+    "prettier-plugin-organize-imports": "4.0.0",
+    "prettier-plugin-tailwindcss": "0.6.5",
+    "tsx": "4.16.5",
+    "typescript": "5.4.2",
+    "vite": "5.4.0",
+    "vite-plugin-dts": "4.0.1"
+  },
+  "peerDependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@fortawesome/free-regular-svg-icons": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
+    "dayjs": "^1.11.12",
+    "formik": "^2.4.6",
+    "lucide-react": "^0.424.0",
+    "postcss": "^8.4.41",
+    "postcss-import": "^16.1.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwind-merge": "^2.4.0",
+    "tailwindcss": "^3.4.7",
+    "tailwindcss-animate": "^1.0.7",
+    "tailwindcss-radix": "^3.0.0",
+    "yup": "^1.0.0"
+  },
+  "scripts": {
+    "build": "run-s build:tsc build:vite build:copy",
+    "build:copy": "tsx src/copy.ts",
+    "build:ladle": "ladle build",
+    "build:tsc": "tsc -b",
+    "build:vite": "vite build",
+    "dev": "ladle serve",
+    "format": "prettier --write .",
+    "format:check": "prettier --check .",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "ladle preview"
+  },
+  "engines": {
+    "node": "=20"
   },
   "volta": {
     "extends": "../../package.json"
-  }
+  },
+  "type": "module"
 }
