@@ -1,36 +1,48 @@
 {
   "name": "@uzh-bf/design-system-custom-element-tag",
-  "version": "0.0.6",
-  "type": "module",
+  "version": "3.0.0-alpha.35",
   "files": [
     "dist"
   ],
-  "source": "src/index.ts",
-  "module": "dist/index.js",
-  "scripts": {
-    "dev": "parcel index.html --open",
-    "build:parcel": "parcel build",
-    "build:purge": "rimraf dist",
-    "build": "run-s build:purge build:parcel"
-  },
   "dependencies": {
-    "@uzh-bf/design-system-react-tag": "*",
+    "@uzh-bf/design-system-react-tag": "workspace:*",
     "htm": "3.1.1",
     "preact": "10.13.2",
     "preact-custom-element": "4.2.1",
     "tailwind-merge": "1.12.0"
   },
   "devDependencies": {
-    "@uzh-bf/parcel-resolver-preact": "*",
-    "@parcel/transformer-inline-string": "2.8.3",
+    "@parcel/plugin": "2.12.0",
+    "@parcel/transformer-inline-string": "2.12.0",
+    "@types/node": "^20.0.0",
     "@types/preact-custom-element": "^4.0.1",
+    "@uzh-bf/parcel-config-design-system": "workspace:*",
+    "@uzh-bf/parcel-resolver-preact": "workspace:*",
     "npm-run-all": "4.1.5",
-    "parcel": "2.8.3",
+    "parcel": "2.12.0",
     "postcss": "8.4.23",
     "rimraf": "5.0.0",
     "tailwindcss": "3.3.2",
-    "typescript": "5.0.4"
+    "typescript": "5.1.6"
   },
+  "scripts": {
+    "build": "run-s build:purge build:parcel",
+    "build:parcel": "parcel build",
+    "build:purge": "rimraf dist",
+    "dev": "parcel index.html --open"
+  },
+  "volta": {
+    "extends": "../../package.json"
+  },
+  "alias": {
+    "react": "preact/compat",
+    "react-dom/test-utils": "preact/test-utils",
+    "react-dom": "preact/compat",
+    "react/jsx-runtime": "preact/jsx-runtime",
+    "react/jsx-dev-runtime": "preact/jsx-runtime"
+  },
+  "module": "dist/index.js",
+  "source": "src/index.ts",
   "targets": {
     "module": {
       "optimize": true,
@@ -43,14 +55,5 @@
       }
     }
   },
-  "alias": {
-    "react": "preact/compat",
-    "react-dom/test-utils": "preact/test-utils",
-    "react-dom": "preact/compat",
-    "react/jsx-runtime": "preact/jsx-runtime",
-    "react/jsx-dev-runtime": "preact/jsx-runtime"
-  },
-  "volta": {
-    "extends": "../../package.json"
-  }
+  "type": "module"
 }
