@@ -1,44 +1,47 @@
 {
   "name": "@uzh-bf/design-system-react-tag",
-  "version": "0.0.4",
-  "type": "module",
+  "version": "3.0.0-alpha.35",
+  "types": "dist/index.d.ts",
   "files": [
     "dist"
   ],
-  "source": "src/Tag.tsx",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "tailwind-merge": "1.12.0"
+  },
   "devDependencies": {
-    "@parcel/packager-ts": "2.8.3",
-    "@parcel/transformer-typescript-types": "2.8.3",
+    "@parcel/packager-ts": "2.12.0",
+    "@parcel/plugin": "2.12.0",
+    "@parcel/transformer-typescript-types": "2.12.0",
+    "@types/node": "^20.0.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.1",
+    "@uzh-bf/parcel-config-design-system": "workspace:*",
     "npm-run-all": "4.1.5",
-    "parcel": "2.8.3",
+    "parcel": "2.12.0",
     "postcss": "8.4.23",
     "rimraf": "5.0.0",
     "tailwindcss": "3.3.2",
-    "typescript": "5.0.4"
+    "typescript": "5.1.6"
+  },
+  "scripts": {
+    "build": "run-s build:purge build:parcel build:tailwind",
+    "build:parcel": "parcel build",
+    "build:purge": "rimraf dist",
+    "build:tailwind": "tailwindcss -i src/styles.css -o dist/styles.css",
+    "dev": "parcel index.html --open"
   },
   "volta": {
     "extends": "../../package.json"
   },
-  "scripts": {
-    "dev": "parcel index.html --open",
-    "build:purge": "rimraf dist",
-    "build:parcel": "parcel build",
-    "build:tailwind": "tailwindcss -i src/styles.css -o dist/styles.css",
-    "build": "run-s build:purge build:parcel build:tailwind"
-  },
-  "dependencies": {
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwind-merge": "1.12.0"
-  },
+  "module": "dist/index.js",
+  "source": "src/Tag.tsx",
   "targets": {
     "module": {
       "optimize": true,
       "includeNodeModules": false
     }
-  }
+  },
+  "type": "module"
 }
