@@ -1,7 +1,7 @@
 {
-  "name": "tc-web",
-  "version": "0.0.6",
-  "description": "Stencil components for TC website",
+  "name": "@uzh-bf/design-system-stencil",
+  "version": "3.0.0-alpha.35",
+  "description": "Web components components for DF",
   "main": "dist/index.cjs.js",
   "module": "./dist/index.js",
   "es2015": "dist/esm/index.mjs",
@@ -21,20 +21,18 @@
     "test.watch": "stencil test --spec --e2e --watchAll",
     "generate": "stencil generate"
   },
-  "dependencies": {
-    "@stencil/core": "^4.18.3",
-    "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/typography": "^0.5.12",
-    "ionicons": "^7.3.1"
-  },
   "devDependencies": {
-    "@stencil/postcss": "^2.1.0",
-    "@stencil/sass": "^2.0.0",
-    "@tailwindcss/forms": "^0.5.3",
-    "autoprefixer": "^10.4.13",
-    "cssnano": "^5.1.14",
-    "stencil-tailwind-plugin": "^1.8.0",
-    "tailwindcss": "^3.2.4",
-    "typescript": "^5.2.2"
+    "@stencil/core": "4.22.3",
+    "@stencil/postcss": "2.1.0",
+    "@stencil/sass": "2.0.0",
+    "@tailwindcss/aspect-ratio": "0.4.2",
+    "@tailwindcss/typography": "0.5.12",
+    "@tailwindcss/forms": "0.5.3",
+    "autoprefixer": "10.4.13",
+    "cssnano": "5.1.14",
+    "ionicons": "7.4.0",
+    "stencil-tailwind-plugin": "1.8.0",
+    "tailwindcss": "3.4.15",
+    "typescript": "5.7.2"
   }
 }
