@@ -1,7 +1,62 @@
-const { VersionConfigTypes } = require('./packages/design-system/src/constants')
-
 module.exports = {
-  types: VersionConfigTypes,
+  types: [
+    {
+      type: 'feat',
+      section: 'Features',
+    },
+    {
+      type: 'enhance',
+      section: 'Enhancements',
+    },
+    {
+      type: 'fix',
+      section: 'Bug Fixes',
+    },
+    {
+      type: 'docs',
+      section: 'Documentation',
+    },
+    {
+      type: 'refactor',
+      section: 'Refactors',
+    },
+    {
+      type: 'perf',
+      section: 'Performance',
+    },
+    {
+      type: 'deploy',
+      section: 'Deployment',
+    },
+    {
+      type: 'deps',
+      section: 'Dependencies',
+    },
+    {
+      type: 'build',
+      section: 'Build and CI',
+    },
+    {
+      type: 'ci',
+      section: 'Build and CI',
+    },
+    {
+      type: 'chore',
+      section: 'Other',
+    },
+    {
+      type: 'wip',
+      section: 'Other',
+    },
+    {
+      type: 'test',
+      section: 'Other',
+    },
+    {
+      type: 'style',
+      section: 'Other',
+    },
+  ],
   packageFiles: [
     {
       filename: `package.json`,
@@ -9,13 +64,19 @@ module.exports = {
     },
   ],
   bumpFiles: [
-    {
-      filename: `package.json`,
+    '',
+    'packages/design-system/',
+    'packages/header-custom-element/',
+    'packages/header-react/',
+    'packages/tag-custom-element/',
+    'packages/tag-react/',
+    'packages/parcel-config-design-system/',
+    'packages/parcel-resolver-preact/',
+    'packages/tailwind-config-design-system/',
+  ].reduce((acc, path) => {
+    return acc.concat({
+      filename: `${path}package.json`,
       type: 'json',
-    },
-    {
-      filename: `package-lock.json`,
-      type: 'json',
-    },
-  ],
+    })
+  }, []),
 }
