@@ -12,7 +12,6 @@ export interface ButtonProps {
   loading?: boolean
   onClick?: (e?: React.MouseEvent<HTMLButtonElement>) => void
   className?: {
-    override?: string
     root?: string
     active?: string
   }
@@ -20,7 +19,7 @@ export interface ButtonProps {
     cy?: string
     test?: string
   }
-  [x: string]: any
+  [x: string]: unknown
 }
 
 /**
@@ -54,7 +53,6 @@ export function Button({
   ...props
 }: ButtonProps) {
   const computedClassName = twMerge(
-    className?.override,
     !basic && 'border rounded px-[0.75em] py-[0.25em] shadow bg-white',
     'inline-flex flex-row items-center font-sans gap-2 cursor-pointer',
     fluid && 'w-full justify-center',
@@ -63,7 +61,7 @@ export function Button({
         ? 'bg-uzh-grey-20 text-uzh-grey-80 cursor-default fill-uzh-grey-80'
         : 'cursor-default'
       : !basic &&
-          `hover:bg-primary-20 hover:border-primary-40 hover:text-primary hover:fill-primary`,
+          `hover:bg-primary-20 hover:border-primary-40 hover:text-primary-100 hover:fill-primary-100`,
     !basic &&
       active &&
       twMerge('bg-primary-20 border-primary-40', className?.active),
@@ -161,7 +159,7 @@ Button.IconGroup = function ButtonIconGroup({
             key={index}
             className={{
               root: twMerge(
-                'p-1.5 first:rounded-l-sm last:rounded-r-sm hover:border-primary',
+                'p-1.5 first:rounded-l-sm last:rounded-r-sm hover:border-primary-100',
                 state === index
                   ? `hover:bg-unset bg-primary-80 text-white`
                   : 'bg-white',
