@@ -1,17 +1,27 @@
-import { faPalette, IconDefinition } from '@fortawesome/free-solid-svg-icons'
+import {
+  faCircleExclamation,
+  faPalette,
+  IconDefinition,
+} from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import React, { useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import { HexColorInput, HexColorPicker } from 'react-colorful'
 import { twMerge } from 'tailwind-merge'
 import Button from './Button'
+import FormLabel from './FormLabel'
 import Label from './forms/Label'
+import { Tooltip } from './Tooltip'
 
 export interface ColorPickerClassName {
   root?: string
+  pickerRoot?: string
+  label?: string
+  tooltip?: string
   trigger?: string
   popover?: string
   presetButtons?: string
   inputLabel?: string
+  inputTooltip?: string
   input?: string
   abort?: string
   submit?: string
@@ -19,22 +29,25 @@ export interface ColorPickerClassName {
 
 export interface ColorPickerProps {
   color: string
+  label?: string
+  labelType?: 'small' | 'large'
+  required?: boolean
   onSubmit: (newColor: string) => void
   disabled?: boolean
   triggerIcon?: IconDefinition
   presetColors?: string[]
-  position?: 'bottom' | 'top'
-  submitText?: string
-  abortText?: string
+  position?: 'bottom' | 'top' | 'bottom-left' | 'top-left'
+  submitText: string
+  colorLabel: string
+  tooltip?: string | React.ReactNode
+  colorTooltip?: string
+  error?: string
+  isTouched?: boolean
   dataTrigger?: {
     cy?: string
     test?: string
   }
   dataHexInput?: {
-    cy?: string
-    test?: string
-  }
-  dataAbort?: {
     cy?: string
     test?: string
   }
@@ -45,18 +58,42 @@ export interface ColorPickerProps {
   className?: ColorPickerClassName
 }
 
+function useOutsideAlerter(elementRef: any, callback: () => void) {
+  useEffect(() => {
+    function handleClickOutside(event: any) {
+      if (elementRef.current && !elementRef.current.contains(event.target)) {
+        callback()
+      }
+    }
+
+    // Bind the event listener
+    document.addEventListener('mousedown', handleClickOutside)
+    return () => {
+      // Unbind the event listener on clean up
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [elementRef])
+}
+
 export function ColorPicker({
   color,
+  label,
+  labelType = 'small',
+  required = false,
   onSubmit,
   disabled,
   triggerIcon,
   presetColors = [],
   position = 'bottom',
   submitText,
-  abortText,
+  colorLabel,
+  tooltip,
+  colorTooltip,
+  error,
+  isTouched,
   dataTrigger,
   dataHexInput,
-  dataAbort,
   dataSubmit,
   className,
 }: ColorPickerProps) {
@@ -67,104 +104,143 @@ export function ColorPicker({
 
   const [colorPickerOpen, setColorPickerOpen] = useState(false)
   const [newColor, setNewColor] = useState(color)
+  const overlayRef = useRef(null)
+  useOutsideAlerter(overlayRef, () => {
+    setColorPickerOpen(false)
+    setNewColor(color)
+  })
 
   return (
     <div
       className={twMerge(
-        'align-center relative flex w-20 justify-end rounded-lg',
+        'flex w-full flex-row',
+        labelType === 'small' && 'flex-col',
         className?.root
       )}
-      style={{ backgroundColor: color ?? '#0028A5' }}
     >
-      <Button
-        onClick={() => setColorPickerOpen(true)}
-        disabled={disabled}
-        data={dataTrigger}
-        className={{ root: className?.trigger }}
-      >
-        <FontAwesomeIcon icon={triggerIcon || faPalette} />
-      </Button>
-      {colorPickerOpen && (
+      {label && (
+        <FormLabel
+          required={required}
+          label={label}
+          labelType={labelType}
+          tooltip={tooltip}
+          className={className}
+        />
+      )}
+      <div className="flex flex-row items-center gap-2">
         <div
           className={twMerge(
-            'absolute flex w-80 flex-col rounded-md bg-white p-1 shadow-md',
-            position === 'bottom' && 'left-10 top-8',
-            position === 'top' && 'bottom-8 left-10',
-            className?.popover
+            'align-center relative flex w-20 justify-end rounded-lg',
+            className?.pickerRoot
           )}
+          style={{ backgroundColor: color ?? '#0028A5' }}
         >
-          <HexColorPicker
-            style={{ width: 'auto' }}
-            color={newColor}
-            onChange={setNewColor}
-          />
-          <div className="grid grid-cols-5 justify-items-center gap-0.5 pb-3 pt-3">
-            {PRESET_COURSE_COLORS.map((presetColor, index) => (
-              <Button
-                key={index}
-                className={{
-                  root: twMerge(
-                    'h-7 w-7 rounded-2xl border-none',
-                    className?.presetButtons
-                  ),
-                }}
-                style={{ backgroundColor: presetColor }}
-                onClick={() => setNewColor(presetColor)}
-              />
-            ))}
-          </div>
-          <div className="flex flex-row">
-            <Label
-              className={{
-                root: twMerge('pr-1', className?.inputLabel),
-                tooltip: 'ml-4',
-              }}
-              label="Farbe:"
-              tooltip="Sie können auch direkt den Hex-Code einer Farbe eingeben"
-              showTooltipSymbol
-            />
-            <HexColorInput
+          <Button
+            onClick={() => setColorPickerOpen(true)}
+            disabled={disabled}
+            data={dataTrigger}
+            className={{ root: className?.trigger }}
+          >
+            <FontAwesomeIcon icon={triggerIcon || faPalette} />
+          </Button>
+          {colorPickerOpen && (
+            <div
               className={twMerge(
-                'focus:border-uzh-blue-50 h-9 w-full rounded border border-uzh-grey-60 pl-2 text-slate-600',
-                className?.input
+                'absolute flex h-[10rem] w-[23rem] flex-row rounded-md bg-white p-1 shadow-md outline outline-2 outline-uzh-grey-40',
+                position === 'bottom' && 'left-10 top-8',
+                position === 'top' && 'bottom-8 left-10',
+                position === 'bottom-left' && '-left-[18rem] top-8',
+                position === 'top-left' && '-left-[18rem] bottom-8',
+                className?.popover
               )}
-              color={newColor}
-              onChange={setNewColor}
-              data-cy={dataHexInput?.cy}
-              data-text={dataHexInput?.test}
-            />
-          </div>
-          <div className="flex w-full flex-row justify-between pt-5">
-            <Button
-              className={{ root: twMerge('float-right', className?.abort) }}
-              type="submit"
-              onClick={() => setColorPickerOpen(false)}
-              data={dataAbort}
+              ref={overlayRef}
             >
-              {abortText ?? 'Abbrechen'}
-            </Button>
-            <Button
-              className={{
-                root: twMerge(
-                  `float-right bg-primary-80 text-white disabled:opacity-60`,
-                  className?.submit
-                ),
-              }}
-              type="submit"
-              onClick={(e) => {
-                e?.preventDefault()
-                e?.stopPropagation()
-                onSubmit(newColor)
-                setColorPickerOpen(false)
-              }}
-              disabled={color === newColor}
-              data={dataSubmit}
-            >
-              {submitText ?? 'Speichern'}
-            </Button>
-          </div>
+              <HexColorPicker
+                style={{ width: '300px', height: '150px' }}
+                color={newColor}
+                onChange={setNewColor}
+              />
+              <div className="grid grid-cols-1 justify-items-center gap-0.5 pl-2">
+                {PRESET_COURSE_COLORS.map((presetColor, index) => (
+                  <Button
+                    key={index}
+                    className={{
+                      root: twMerge(
+                        'h-7 w-7 rounded-2xl border-none',
+                        className?.presetButtons
+                      ),
+                    }}
+                    style={{ backgroundColor: presetColor }}
+                    onClick={() => setNewColor(presetColor)}
+                  />
+                ))}
+              </div>
+
+              <div className="ml-4 flex flex-col justify-between">
+                <div className="flex flex-col">
+                  <Label
+                    label={colorLabel}
+                    className={{
+                      root: twMerge(
+                        'my-auto -mb-0.5 min-w-max text-base font-bold leading-6 text-gray-600',
+                        className?.inputLabel
+                      ),
+                      tooltip: twMerge(
+                        'text-sm font-normal',
+                        className?.inputTooltip
+                      ),
+                      tooltipSymbol: 'h-2 w-2',
+                    }}
+                    tooltip={colorTooltip}
+                    showTooltipSymbol={typeof colorTooltip !== 'undefined'}
+                  />
+                  <HexColorInput
+                    className={twMerge(
+                      'focus:border-uzh-blue-50 h-9 w-24 rounded border border-uzh-grey-60 pl-2 placeholder-slate-400',
+                      className?.input
+                    )}
+                    color={newColor}
+                    onChange={setNewColor}
+                    data-cy={dataHexInput?.cy}
+                    data-text={dataHexInput?.test}
+                  />
+                </div>
+                <Button
+                  className={{
+                    root: twMerge(
+                      `float-right h-max w-full justify-center self-end bg-primary-80 text-white disabled:opacity-60`,
+                      className?.submit
+                    ),
+                  }}
+                  type="submit"
+                  onClick={(e) => {
+                    e?.preventDefault()
+                    e?.stopPropagation()
+                    onSubmit(newColor)
+                    setColorPickerOpen(false)
+                  }}
+                  disabled={color === newColor}
+                  data={dataSubmit}
+                >
+                  {submitText}
+                </Button>
+              </div>
+            </div>
+          )}
         </div>
-      )}
+        {error && isTouched && (
+          <Tooltip
+            tooltip={error}
+            delay={0}
+            className={{ tooltip: 'max-w-[30rem] text-sm' }}
+          >
+            <FontAwesomeIcon
+              icon={faCircleExclamation}
+              className="mr-1 text-red-600"
+            />
+          </Tooltip>
+        )}
+      </div>
     </div>
   )
 }
