@@ -36,8 +36,6 @@ export interface ModalProps {
     test?: string
   }
   className?: {
-    overlayOverride?: string
-    contentOverride?: string
     overlay?: string
     content?: string
     title?: string
@@ -131,7 +129,6 @@ export function Modal({
   const overlayContent = (
     <RadixDialog.Overlay
       className={twMerge(
-        className?.overlayOverride,
         'fixed bottom-0 left-0 right-0 top-0 z-20 flex justify-center gap-4 bg-uzh-grey-100 bg-opacity-50 p-4 md:items-center',
         className?.overlay
       )}
@@ -151,11 +148,10 @@ export function Modal({
 
       <RadixDialog.Content
         className={twMerge(
-          className?.contentOverride,
           'z-30 flex flex-col gap-4 overflow-y-auto rounded-lg bg-white p-4 shadow',
           fullScreen
             ? 'h-full w-full'
-            : 'min-h-[18rem] w-[27rem] max-w-7xl md:h-[28rem] md:w-[40rem] lg:h-[40rem] lg:w-[55rem] xl:h-[45rem] xl:w-[70rem]',
+            : 'h-max max-h-full w-[27rem] max-w-7xl md:w-[40rem] lg:w-[55rem] xl:w-[70rem]',
           className?.content
         )}
         onEscapeKeyDown={escapeDisabled ? undefined : onClose}
@@ -194,10 +190,12 @@ export function Modal({
 
         <div className={twMerge('flex-1', className?.children)}>{children}</div>
 
-        <div className="flex flex-initial flex-row justify-between">
-          <div>{onSecondaryAction && <div>{onSecondaryAction}</div>}</div>
-          <div>{onPrimaryAction && <div>{onPrimaryAction}</div>}</div>
-        </div>
+        {(onPrimaryAction || onSecondaryAction) && (
+          <div className="flex flex-initial flex-row justify-between">
+            <div>{onSecondaryAction && <div>{onSecondaryAction}</div>}</div>
+            <div>{onPrimaryAction && <div>{onPrimaryAction}</div>}</div>
+          </div>
+        )}
       </RadixDialog.Content>
 
       {(onPrev || onNext) && (
