@@ -10,7 +10,6 @@ export interface NavigationProps {
   }
   children: React.ReactNode
   className?: {
-    override?: string
     root?: string
     indicator?: string
     viewport?: string
@@ -42,7 +41,6 @@ export function Navigation({
   return (
     <NavigationMenuPrimitive.Root
       className={twMerge(
-        className?.override,
         'relative w-max rounded-md bg-primary-20',
         className?.root
       )}
@@ -86,8 +84,6 @@ interface TriggerProps {
   children: React.ReactNode
   disabled?: boolean
   className?: {
-    override?: string
-    dropdownOverride?: string
     root?: string
     label?: string
     icon?: string
@@ -144,7 +140,6 @@ Navigation.TriggerItem = function TriggerItem({
         data-cy={data?.cy}
         data-test={data?.test}
         className={twMerge(
-          className?.override,
           'flex flex-row items-center rounded-md px-3 py-2 text-sm font-medium text-black hover:text-white focus:outline-none focus-visible:ring',
           icon && !label && 'h-9 w-9 justify-center',
           !disabled && 'hover:bg-primary-60',
@@ -177,7 +172,6 @@ Navigation.TriggerItem = function TriggerItem({
 
       <NavigationMenuPrimitive.Content
         className={twMerge(
-          className?.dropdownOverride,
           'flex flex-col gap-2 rounded-lg p-2',
           dropdownWidth,
           className?.dropdown
@@ -200,7 +194,6 @@ interface DropdownItemProps {
   subtitle?: string
   icon?: React.ReactNode
   className?: {
-    override?: string
     root?: string
     title?: string
     icon?: string
@@ -258,9 +251,8 @@ Navigation.DropdownItem = function DropdownItem({
       href={href}
       onClick={onClick}
       className={twMerge(
-        className?.override,
         'w-full rounded-md px-4 py-3 text-black hover:bg-primary-60',
-        'focus:outline-none focus-visible:ring focus-visible:ring-opacity-75 sm:hover:text-white',
+        'hover:text-white focus:outline-none focus-visible:ring focus-visible:ring-opacity-75',
         className?.root
       )}
       style={style?.root}
@@ -308,7 +300,6 @@ interface ButtonItemProps {
   disabled?: boolean
   icon?: React.ReactNode
   className?: {
-    override?: string
     root?: string
     label?: string
     icon?: string
@@ -366,8 +357,7 @@ Navigation.ButtonItem = function ButtonItem({
         href={!disabled ? href : undefined}
         onClick={!disabled ? onClick : undefined}
         className={twMerge(
-          className?.override,
-          'cursor-pointer rounded-md px-3 py-2 text-sm font-medium text-black sm:hover:text-white',
+          'cursor-pointer rounded-md px-3 py-2 text-sm font-medium text-black hover:text-white',
           !disabled && 'hover:bg-primary-60',
           disabled && 'hover:text-none cursor-not-allowed text-gray-400',
           className?.root,
@@ -402,7 +392,6 @@ interface IconItemProps {
   icon: React.ReactNode
   disabled?: boolean
   className?: {
-    override?: string
     root?: string
     disabled?: string
   }
@@ -454,8 +443,7 @@ Navigation.IconItem = function IconItem({
         href={!disabled ? href : undefined}
         onClick={!disabled ? onClick : undefined}
         className={twMerge(
-          className?.override,
-          'flex h-9 w-9 items-center justify-center rounded-md text-black sm:hover:text-white',
+          'flex h-9 w-9 items-center justify-center rounded-md text-black hover:text-white',
           !disabled && 'hover:bg-primary-60',
           disabled && 'hover:text-none cursor-not-allowed text-gray-400',
           className?.root,
