@@ -1,51 +1,12 @@
-import React, { useState } from 'react'
+import { useState } from 'react'
 import Select from './Select'
-
-const fruits = [
-  { value: 'apple', label: 'Apple' },
-  { value: 'banana', label: 'Banana' },
-  { value: 'pear', label: 'Pear' },
-  { value: 'watermeldon', label: 'Watermelon' },
-  { value: 'peach', label: 'Peach' },
-  { value: 'mango', label: 'Mango' },
-]
-const vegetables = [
-  { value: 'carrot', label: 'Carrot' },
-  { value: 'cucumber', label: 'Cucumber' },
-  { value: 'onion', label: 'Onion' },
-  { value: 'potato', label: 'Potato' },
-  { value: 'tomato', label: 'Tomato' },
-  { value: 'broccoli', label: 'Broccoli' },
-]
-const transport = [
-  { value: 'car', label: 'Car' },
-  { value: 'bike', label: 'Bike' },
-  { value: 'train', label: 'Train' },
-  { value: 'plane', label: 'Plane' },
-  { value: 'boat', label: 'Boat' },
-  { value: 'bus', label: 'Bus' },
-]
-const programming = [
-  { value: 'javascript', label: 'JavaScript' },
-  { value: 'typescript', label: 'TypeScript' },
-  { value: 'python', label: 'Python' },
-  { value: 'ruby', label: 'Ruby' },
-  { value: 'java', label: 'Java' },
-  { value: 'csharp', label: 'C#' },
-]
-
-const groups = [
-  { items: fruits },
-  { items: vegetables, showSeparator: true },
-  { items: transport, showSeparator: true, label: 'Transport' },
-  { items: programming, showSeparator: true, label: 'Programming Languages' },
-]
+import { fruitsValues, groupValues } from './values'
 
 export const Default = () => {
   return (
     <Select
       placeholder="Select an item"
-      items={fruits}
+      items={fruitsValues}
       onChange={(newValue) => {
         console.log(newValue)
       }}
@@ -58,7 +19,7 @@ export const Popper = () => {
     <Select
       contentPosition="popper"
       placeholder="Select an item"
-      items={fruits}
+      items={fruitsValues}
       onChange={(newValue) => {
         console.log(newValue)
       }}
@@ -70,7 +31,7 @@ export const Basic = () => {
   return (
     <Select
       placeholder="Select an item"
-      items={fruits}
+      items={fruitsValues}
       onChange={(newValue) => {
         console.log(newValue)
       }}
@@ -83,11 +44,11 @@ export const DefaultValue = () => {
   return (
     <Select
       placeholder="Select an item"
-      items={fruits}
+      items={fruitsValues}
       onChange={(newValue) => {
         console.log(newValue)
       }}
-      defaultValue={fruits[2].value}
+      defaultValue={fruitsValues[2].value}
     />
   )
 }
@@ -106,7 +67,7 @@ export const Groups = () => {
       <Select
         placeholder="Select an item"
         groups={[
-          ...groups,
+          ...groupValues,
           {
             label: 'Short labels',
             items: [
@@ -126,7 +87,7 @@ export const Groups = () => {
                 label: 'Pear long label',
                 shortLabel: 'Pear',
               },
-              { value: 'watermeldon_short', label: 'Watermelon' },
+              { value: 'watermelon_short', label: 'Watermelon' },
             ],
           },
         ]}
@@ -160,7 +121,7 @@ export const DisabledElements = () => {
     { value: 'apple', label: 'Apple' },
     { value: 'banana', label: 'Banana', disabled: true },
     { value: 'pear', label: 'Pear' },
-    { value: 'watermeldon', label: 'Watermelon' },
+    { value: 'watermelon', label: 'Watermelon' },
     { value: 'peach', label: 'Peach', disabled: true },
     { value: 'mango', label: 'Mango' },
   ]
@@ -181,16 +142,15 @@ export const Styled = () => {
     <div className="h-12 w-full rounded-md border border-solid">
       <Select
         placeholder="Select an item"
-        groups={groups}
+        groups={groupValues}
         onChange={(newValue) => {
           console.log(newValue)
         }}
         className={{
           root: 'h-full rounded-none border-none',
           content: 'bg-green-200',
-          trigger: 'rounded-none border-none bg-uzh-blue-20',
+          trigger: 'rounded-none border-none bg-uzh-blue-20 text-lg',
           item: 'text-red-700',
-          scrollButton: 'bg-blue-400',
           groupLabel: 'text-red-700',
           separator: 'bg-green-700',
         }}
@@ -199,23 +159,12 @@ export const Styled = () => {
   )
 }
 
-export const Small = () => {
-  return (
-    <Select
-      placeholder="Select an item"
-      groups={groups}
-      onChange={(newValue) => console.log(newValue)}
-      size="sm"
-    />
-  )
-}
-
 export const CustomTriggerLabel = () => {
   const items = [
     { value: 'apple', label: 'Apple long label', shortLabel: 'Apple' },
     { value: 'banana', label: 'Banana long label', shortLabel: 'Banana' },
     { value: 'pear', label: 'Pear long label' },
-    { value: 'watermeldon', label: 'Watermelon long label' },
+    { value: 'watermelon', label: 'Watermelon long label' },
     { value: 'peach', label: 'Peach long label' },
     { value: 'mango', label: 'Mango long label', shortLabel: 'Mango' },
   ]
