@@ -1,23 +1,21 @@
-import {
-  faCheck,
-  faChevronDown,
-  faChevronUp,
-} from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import * as RadixSelect from '@radix-ui/react-select'
-import React, { useState } from 'react'
+import { useEffect, useState } from 'react'
 import { twMerge } from 'tailwind-merge'
+import {
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectSeparator,
+  SelectTrigger,
+  SelectValue,
+  Select as ShadcnSelect,
+} from './ui/select'
 
 export interface SelectClassName {
-  triggerOverride?: string
-  contentOverride?: string
-  itemOverride?: string
   root?: string
   trigger?: string
   content?: string
   item?: string
-  text?: string
-  scrollButton?: string
   groupLabel?: string
   separator?: string
 }
@@ -33,16 +31,14 @@ interface SelectProps {
   onBlur?: () => void
   value?: string
   disabled?: boolean
-  size?: 'md' | 'sm'
   className?: SelectClassName
   placeholder?: string
   defaultValue?: string
   basic?: boolean
-  asPortal?: boolean
   contentPosition?: 'item-aligned' | 'popper'
 }
 
-export interface Item {
+export interface SelectItem {
   id?: string
   data?: {
     cy?: string
@@ -54,17 +50,19 @@ export interface Item {
   shortLabel?: string
 }
 
+export interface SelectGroup {
+  label?: string
+  showSeparator?: boolean
+  items: SelectItem[]
+}
+
 export interface SelectWithItemsProps extends SelectProps {
-  items: Item[]
+  items: SelectItem[]
   groups?: never
 }
 
 export interface SelectWithGroupsProps extends SelectProps {
-  groups: {
-    label?: string
-    showSeparator?: boolean
-    items: Item[]
-  }[]
+  groups: SelectGroup[]
   items?: never
 }
 
@@ -83,10 +81,8 @@ export interface SelectWithGroupsProps extends SelectProps {
  * @param defaultValue - The default value of the select component set initially.
  * @param placeholder - The placeholder text that is displayed when no value is selected.
  * @param disabled - Specifies whether the select component is disabled or not.
- * @param size - The size of the select component. Currently only medium and small are supported.
  * @param basic - Specifies whether the select component is basic or not. A basic select component does only have minimal styling of the trigger.
  * @param className - The optional className object allows you to override the default styling.
- * @param asPortal - If true, the select component is rendered as a portal.
  * @param contentPosition - The position of the content of the select component. Currently only 'item-aligned' and 'popper' are supported.
  * @return Select component
  */
@@ -99,77 +95,26 @@ export function Select({
   onBlur,
   value,
   disabled = false,
-  size = 'md',
   className,
   name,
   placeholder,
   defaultValue,
   basic = false,
-  asPortal = false,
   contentPosition = 'item-aligned',
 }: SelectWithItemsProps | SelectWithGroupsProps) {
   const [open, setOpen] = useState(false)
-  const flatItems = items || groups?.flatMap((group) => group.items) || []
+  const [shortLabel, setShortLabel] = useState<string | undefined>(undefined)
 
-  const selectContent = (
-    <RadixSelect.Content
-      position={contentPosition}
-      className={twMerge(
-        className?.contentOverride,
-        'z-50 overflow-hidden rounded-md bg-white shadow-md',
-        className?.content
-      )}
-    >
-      <RadixSelect.ScrollUpButton
-        className={twMerge(
-          'flex h-7 items-center justify-center bg-white',
-          className?.scrollButton
-        )}
-      >
-        <FontAwesomeIcon
-          icon={faChevronUp}
-          size={size === 'sm' ? 'sm' : '1x'}
-        />
-      </RadixSelect.ScrollUpButton>
-      <RadixSelect.Viewport className="rounded-lg p-1">
-        {items
-          ? items.map((item, ix) => (
-              <SelectItem
-                id={id}
-                data={item.data}
-                key={ix}
-                size={size}
-                {...item}
-                className={className}
-              />
-            ))
-          : groups.map((group, ix) => (
-              <SelectGroup
-                key={ix}
-                size={size}
-                {...group}
-                className={className}
-              />
-            ))}
-      </RadixSelect.Viewport>
-      <RadixSelect.ScrollDownButton
-        className={twMerge(
-          'flex h-7 items-center justify-center bg-white',
-          className?.scrollButton
-        )}
-      >
-        <FontAwesomeIcon
-          icon={faChevronDown}
-          size={size === 'sm' ? 'sm' : '1x'}
-        />
-      </RadixSelect.ScrollDownButton>
-    </RadixSelect.Content>
-  )
+  useEffect(() => {
+    const flatItems = items || groups?.flatMap((group) => group.items) || []
+    setShortLabel(flatItems?.find((item) => item.value === value)?.shortLabel)
+  }, [value, items, groups])
 
   return (
     <div className={twMerge('relative flex', className?.root)}>
-      <RadixSelect.Root
+      <ShadcnSelect
         name={name}
+        open={open}
         onValueChange={onChange}
         onOpenChange={(open) => {
           setOpen(open)
@@ -180,152 +125,81 @@ export function Select({
         value={value}
         defaultValue={defaultValue}
       >
-        <RadixSelect.Trigger
+        <SelectTrigger
           id={id}
           data-cy={data?.cy}
           data-test={data?.test}
           className={twMerge(
-            className?.triggerOverride,
-            'rounded-md px-2 py-1',
-            !basic &&
-              'inline-flex h-7 items-center justify-between gap-2 border  bg-white p-4 shadow-sm hover:bg-primary-20 sm:hover:text-primary',
-            disabled &&
-              'hover:bg-none, hover:text-none cursor-not-allowed bg-uzh-grey-20 opacity-70 shadow-sm',
-            size === 'sm' && '!text-sm',
+            'w-[15rem] text-base',
+            disabled && 'bg-uzh-grey-20 opacity-70',
+            basic && '[all:_unset]',
             className?.trigger
           )}
+          hideIcon={basic}
           disabled={disabled}
         >
-          <div
-            className={
-              flatItems?.find((item) => item.value === value)?.shortLabel &&
-              'hidden'
-            }
-          >
-            <RadixSelect.Value placeholder={placeholder} />
-          </div>
-          <div
-            className={twMerge(
-              'hidden',
-              flatItems?.find((item) => item.value === value)?.shortLabel &&
-                'block'
-            )}
-          >
-            {flatItems?.find((item) => item.value === value)?.shortLabel}
-          </div>
-
-          {!basic && (
-            <RadixSelect.Icon
-              className={twMerge('ml-2', size === 'sm' && 'ml-0.5')}
-            >
-              <FontAwesomeIcon
-                icon={open ? faChevronUp : faChevronDown}
-                size={size === 'sm' ? 'sm' : '1x'}
-              />
-            </RadixSelect.Icon>
+          {shortLabel ? (
+            <SelectValue placeholder={placeholder} key={shortLabel}>
+              {shortLabel}
+            </SelectValue>
+          ) : (
+            <SelectValue placeholder={placeholder} key={shortLabel} />
           )}
-        </RadixSelect.Trigger>
-        {asPortal ? (
-          <RadixSelect.Portal>{selectContent}</RadixSelect.Portal>
-        ) : (
-          selectContent
-        )}
-      </RadixSelect.Root>
-    </div>
-  )
-}
-
-interface SelectItemProps {
-  id?: string
-  data?: {
-    cy?: string
-    test?: string
-  }
-  label: string
-  className?: SelectClassName
-  size?: string
-  value: string
-  disabled?: boolean
-}
-
-const SelectItem = React.forwardRef(
-  (
-    { id, data, className, label, size, disabled, ...props }: SelectItemProps,
-    forwardedRef
-  ) => {
-    return (
-      <RadixSelect.Item
-        id={id}
-        data-cy={data?.cy}
-        data-test={data?.test}
-        className={twMerge(
-          className?.itemOverride,
-          'relative flex select-none items-center rounded-md px-8 py-2 font-medium text-gray-700',
-          'hover:cursor-pointer hover:bg-primary-20 hover:outline-none focus:border-primary-40 sm:hover:text-primary',
-          disabled &&
-            'cursor-not-allowed opacity-50 hover:bg-white hover:text-gray-700',
-          size === 'sm' && 'px-7 text-sm',
-          className?.item
-        )}
-        disabled={disabled}
-        {...props}
-        ref={forwardedRef as React.ForwardedRef<HTMLDivElement>}
-      >
-        <RadixSelect.ItemText className={twMerge(className?.text)}>
-          {label}
-        </RadixSelect.ItemText>
-        <RadixSelect.ItemIndicator className="absolute left-2 inline-flex items-center">
-          <FontAwesomeIcon icon={faCheck} size={size === 'sm' ? 'sm' : '1x'} />
-        </RadixSelect.ItemIndicator>
-      </RadixSelect.Item>
-    )
-  }
-)
-
-interface SelectGroupProps {
-  items: Item[]
-  size?: 'md' | 'sm'
-  showSeparator?: boolean
-  label?: string
-  className?: SelectClassName
-}
-
-const SelectGroup = ({
-  items,
-  size,
-  showSeparator,
-  label,
-  className,
-  ...props
-}: SelectGroupProps) => {
-  return (
-    <>
-      {showSeparator && (
-        <RadixSelect.Separator
-          className={twMerge('h-0.5 bg-black', className?.separator)}
-        />
-      )}
-      <RadixSelect.Group {...props}>
-        <RadixSelect.Label
-          className={twMerge(
-            'p-1 font-bold',
-            size === 'sm' && '!text-sm',
-            className?.groupLabel
-          )}
+        </SelectTrigger>
+        <SelectContent
+          position={contentPosition}
+          className={twMerge('z-50', className?.content)}
         >
-          {label}
-        </RadixSelect.Label>
-        {items.map((item, ix) => (
-          <SelectItem
-            key={ix}
-            size={size}
-            data={item.data}
-            {...item}
-            className={className}
-          />
-        ))}
-      </RadixSelect.Group>
-    </>
+          {items
+            ? items.map((item, ix) => (
+                <SelectItem
+                  id={id}
+                  key={`${item.value}-${ix}`}
+                  value={item.value}
+                  disabled={item.disabled}
+                  data-cy={item.data?.cy}
+                  data-test={item.data?.test}
+                  className={twMerge(
+                    'text-base',
+                    item.disabled && 'cursor-not-allowed',
+                    className?.item
+                  )}
+                >
+                  {item.label}
+                </SelectItem>
+              ))
+            : groups.map((group, ix) => (
+                <SelectGroup key={`group-${group.label}-${ix}`}>
+                  {group.showSeparator && <SelectSeparator />}
+                  {group.label && (
+                    <SelectLabel
+                      className={twMerge('text-base', className?.groupLabel)}
+                    >
+                      {group.label}
+                    </SelectLabel>
+                  )}
+                  {group.items.map((item, item_ix) => (
+                    <SelectItem
+                      id={id}
+                      key={`${item.value}-${item_ix}`}
+                      value={item.value}
+                      disabled={item.disabled}
+                      data-cy={item.data?.cy}
+                      data-test={item.data?.test}
+                      className={twMerge(
+                        'text-base',
+                        item.disabled && '!cursor-not-allowed',
+                        className?.item
+                      )}
+                    >
+                      {item.label}
+                    </SelectItem>
+                  ))}
+                </SelectGroup>
+              ))}
+        </SelectContent>
+      </ShadcnSelect>
+    </div>
   )
 }
 
