@@ -14,7 +14,6 @@ interface TabProps {
   children?: React.ReactNode
   disabled?: boolean
   className?: {
-    override?: string
     root?: string
     label?: string
     disabled?: string
@@ -63,7 +62,6 @@ export function Tab({
       key={`tab-trigger-${key}`}
       value={value}
       className={twMerge(
-        className?.override,
         'group flex-1 border-b border-r border-gray-300 px-3 py-2.5 first:rounded-tl-lg last:rounded-tr-lg last:border-r-0',
         'rdx-state-active:border-b-slate-600 focus-visible:rdx-state-active:border-b-transparent rdx-state-inactive:bg-gray-50 hover:rdx-state-inactive:bg-gray-200',
         'focus:rdx-state-active:border-b-red focus:z-10 focus:outline-none focus-visible:ring focus-visible:ring-purple-500 focus-visible:ring-opacity-75',
@@ -90,7 +88,6 @@ interface TabListProps {
     test?: string
   }
   className?: {
-    override?: string
     root?: string
   }
 }
@@ -116,7 +113,6 @@ export function TabList({
       data-cy={data?.cy}
       data-test={data?.test}
       className={twMerge(
-        className?.override,
         'flex w-full flex-col rounded-t-lg bg-white md:flex-row',
         className?.root
       )}
@@ -135,7 +131,6 @@ interface TabContentProps {
   key: string
   value: string
   className?: {
-    override?: string
     root?: string
   }
 }
@@ -167,11 +162,7 @@ export function TabContent({
       data-test={data?.test}
       key={`tab-content-${key}`}
       value={value}
-      className={twMerge(
-        className?.override,
-        'rounded-t-lg bg-white py-4 md:px-6',
-        className?.root
-      )}
+      className={twMerge('rounded-t-lg bg-white py-4 md:px-6', className?.root)}
     >
       {children}
     </TabsPrimitive.Content>
