@@ -1,23 +1,27 @@
+import { faCircleExclamation } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useField } from 'formik'
 import React from 'react'
 import { twMerge } from 'tailwind-merge'
-import Label from './Label'
+import FormLabel from '../FormLabel'
+import { Tooltip } from '../Tooltip'
 
-export interface DateFieldProps {
+export interface FormikDateFieldProps {
   id?: string
+  name: string
   data?: {
     cy?: string
     test?: string
   }
   label?: string
-  labelType?: 'small' | 'normal'
+  labelType?: 'small' | 'large'
   placeholder?: string
   tooltip?: string | React.ReactNode
   required?: boolean
   hideError?: boolean
+  touchedOnChange?: boolean
   disabled?: boolean
   className?: {
-    override?: string
     root?: string
     field?: string
     label?: string
@@ -25,8 +29,7 @@ export interface DateFieldProps {
     error?: string
     tooltip?: string
   }
-  name: string
-  [key: string]: any
+  [key: string]: unknown
 }
 
 /**
@@ -42,6 +45,7 @@ export interface DateFieldProps {
  * @param tooltip - The optional tooltip is shown on hover next to the label.
  * @param required - Indicate whether the field is required or not.
  * @param hideError - Hide the error message below this component as is might be more appropriate to show it somewhere else.
+ * @param touchedOnChange - Indicate whether the field should be marked as touched on change.
  * @param disabled - Disable the field.
  * @param className - The optional className object allows you to override the default styling.
  * @returns Date field component with Formik state management.
@@ -51,19 +55,20 @@ export function FormikDateField({
   data,
   name,
   label,
-  labelType,
+  labelType = 'small',
   placeholder,
   tooltip,
   required = false,
   hideError = false,
+  touchedOnChange = false,
   disabled = false,
   className,
   ...props
-}: DateFieldProps) {
+}: FormikDateFieldProps) {
   const [field, meta, helpers] = useField(name)
 
   return (
-    <div className={twMerge('flex flex-col', className?.root)}>
+    <div className={twMerge('w-max', className?.root)}>
       <div
         className={twMerge(
           'flex w-full flex-row',
@@ -72,59 +77,56 @@ export function FormikDateField({
         )}
       >
         {label && (
-          <Label
-            forId={id}
+          <FormLabel
+            id={id}
             required={required}
             label={label}
-            className={{
-              root: twMerge(
-                'my-auto mr-2 min-w-max font-bold',
-                labelType === 'small' &&
-                  'mt-1 text-sm font-normal leading-6 text-gray-600',
-                className?.label
-              ),
-              tooltip: twMerge('text-sm font-normal', className?.tooltip),
-              tooltipSymbol: twMerge(labelType === 'small' && 'h-2 w-2'),
-            }}
+            labelType={labelType}
             tooltip={tooltip}
-            showTooltipSymbol={typeof tooltip !== 'undefined'}
+            className={className}
           />
         )}
+        <div className="flex flex-row items-center gap-2">
+          <input
+            id={id}
+            data-cy={data?.cy}
+            data-test={data?.test}
+            type="datetime-local"
+            value={field.value}
+            onChange={(e) => {
+              if (e.target['validity'].valid) {
+                helpers.setValue(e.target['value'])
 
-        <input
-          id={id}
-          data-cy={data?.cy}
-          data-test={data?.test}
-          type="datetime-local"
-          value={field.value}
-          onChange={(e) => {
-            if (e.target['validity'].valid) {
-              helpers.setValue(e.target['value'])
-            }
-          }}
-          onBlur={() => helpers.setTouched(true)}
-          placeholder={placeholder}
-          disabled={disabled}
-          className={twMerge(
-            className?.override,
-            'focus:border-uzh-blue-50 h-9 w-full rounded border border-uzh-grey-60 bg-uzh-grey-20',
-            disabled && 'cursor-not-allowed text-uzh-grey-100',
-            meta.error && meta.touched && 'border-red-400 bg-red-50',
-            className?.input
+                if (touchedOnChange) {
+                  helpers.setTouched(true)
+                }
+              }
+            }}
+            onBlur={() => helpers.setTouched(true)}
+            placeholder={placeholder}
+            disabled={disabled}
+            className={twMerge(
+              'focus:border-uzh-blue-50 h-9 w-full rounded border border-uzh-grey-60 bg-uzh-grey-20',
+              disabled && 'cursor-not-allowed text-uzh-grey-100',
+              meta.error && meta.touched && 'border-red-400 bg-red-50',
+              className?.input
+            )}
+            {...props}
+          />
+          {meta.error && !hideError && meta.touched && (
+            <Tooltip
+              tooltip={meta.error}
+              delay={0}
+              className={{ tooltip: 'text ml-3 max-w-[30rem]' }}
+            >
+              <FontAwesomeIcon
+                icon={faCircleExclamation}
+                className="mr-1 text-red-600"
+              />
+            </Tooltip>
           )}
-          {...props}
-        />
-      </div>
-      {!hideError && meta.error && (
-        <div
-          className={twMerge(
-            'w-full text-right text-sm text-red-400',
-            className?.error
-          )}
-        >
-          {meta.error}
         </div>
-      )}
+      </div>
     </div>
   )
 }
