@@ -1,136 +1,147 @@
 import { useField } from 'formik'
 import React from 'react'
 import { twMerge } from 'tailwind-merge'
-import Label from './Label'
 import NumberField, { NumberFieldClassName } from './NumberField'
 
-export interface FormikNumberFieldProps {
+interface FormikNumberFieldProps {
   id?: string
-  name: string
+  label?: string
+  labelType?: 'small' | 'large'
+  placeholder?: string
+  precision?: number
+  min?: number
+  max?: number
+  tooltip?: string | React.ReactNode
+  required?: boolean
+  hideError?: boolean
+  error?: string
+  disabled?: boolean
+  onBlur?: () => void
   data?: {
     cy?: string
     test?: string
   }
-  label?: string
-  labelType?: 'small' | 'normal'
-  placeholder?: string
-  tooltip?: string | React.ReactNode
-  required?: boolean
-  hideError?: boolean
-  precision?: number
-  min?: number
-  max?: number
-  disabled?: boolean
-  className?: {
-    root?: string
-    field?: string
-    label?: string
-    tooltip?: string
-    error?: string
-    numberField?: NumberFieldClassName
-  }
+  className?: NumberFieldClassName & { root?: string }
+  [key: string]: unknown
+}
+
+export interface FormikNumberFieldNameProps extends FormikNumberFieldProps {
+  name: string
+  value?: never
+  onChange?: never
+  isTouched?: never
+}
+
+export interface FormikNumberFieldOnChangeProps extends FormikNumberFieldProps {
+  name?: never
+  value: string
+  onChange: (newValue: string) => void
+  isTouched?: boolean
 }
 
 /**
- * This function returns a text field that works as to be expected in a Formik environment.
- * State can be managed either through Formik or internally by passing a value and onChange function.
+ * This function returns a text field component for use without formik
  *
- * @param id - The id of the field.
- * @param data - The object of data attributes that can be used for testing (e.g. data-test or data-cy)
- * @param name - The name of the field as used to keep track of the state in Formik. If no value and onChange function are provided, this field is required.
- * @param label - The optional label is shown next to the field in the form.
+ * @param id - The id of the input field.
+ * @param value - The value of the input field (external state management).
+ * @param onChange - The onChange function of the input field (external state management).
+ * @param label - The text displayed as label.
  * @param labelType - The optional labelType can be used to change the size and position of the label according to pre-defined standards.
- * @param placeholder - The optional placeholder is shown when the field is empty.
- * @param tooltip - The optional tooltip is shown on hover next to the label.
- * @param required - Indicate whether the field is required or not.
- * @param hideError - Hide the error message below this component as is might be more appropriate to show it somewhere else.
+ * @param placeholder - The placeholder text for the input field.
  * @param precision - The optional precision defines the number of decimal places that are allowed.
  * @param min - The optional min defines the minimum value that is allowed.
  * @param max - The optional max defines the maximum value that is allowed.
- * @param disabled - Disables the field.
+ * @param tooltip - The optional tooltip is shown on hover over the tooltip next to the label.
+ * @param required - Indicate whether the field is required or not.
+ * @param hideError - Indicate whether the error message should be hidden or not.
+ * @param error - The error message that is displayed below the input field.
+ * @param isTouched - Indicate whether the field has been touched or not (validation is not handled by this component).
+ * @param disabled - Indicate whether the field is disabled or not.
+ * @param onBlur - The onBlur function of the input field.
+ * @param data - The object of data attributes that can be used for testing (e.g. data-test or data-cy)
  * @param className - The optional className object allows you to override the default styling.
- * @returns Text field component with Formik state management.
  */
 export function FormikNumberField({
   id,
-  data,
   name,
+  value,
+  onChange,
   label,
-  labelType,
+  labelType = 'small',
   placeholder,
-  tooltip,
-  required = false,
-  hideError = false,
   precision,
   min,
   max,
-  disabled = false,
+  tooltip,
+  required,
+  hideError,
+  error,
+  isTouched,
+  disabled,
+  onBlur,
+  data,
   className,
-}: FormikNumberFieldProps) {
-  const [field, meta, helpers] = useField(name)
+  ...props
+}: FormikNumberFieldNameProps | FormikNumberFieldOnChangeProps) {
+  const [field, meta, helpers] = useField(name || '')
 
-  return (
-    <div className={twMerge('flex flex-col', className?.root)}>
-      <div
-        className={twMerge(
-          'flex w-full flex-row',
-          labelType === 'small' && 'flex-col',
-          className?.field
-        )}
-      >
-        {label && (
-          <Label
-            forId={id}
-            required={required}
-            label={label}
-            className={{
-              root: twMerge(
-                'my-auto mr-2 min-w-max font-bold',
-                labelType === 'small' &&
-                  'mt-1 text-sm font-normal leading-6 text-gray-600',
-                className?.label
-              ),
-              tooltip: twMerge('text-sm font-normal', className?.tooltip),
-              tooltipSymbol: twMerge(labelType === 'small' && 'h-2 w-2'),
-            }}
-            tooltip={tooltip}
-            showTooltipSymbol={typeof tooltip !== 'undefined'}
-          />
-        )}
+  if (name) {
+    return (
+      <div className={twMerge('w-full', className?.root)}>
         <NumberField
           id={id}
-          data={data}
           value={field.value}
-          onChange={(newValue) => helpers.setValue(newValue)}
-          onBlur={() => helpers.setTouched(true)}
+          onChange={(newValue: string) => helpers.setValue(newValue)}
+          label={label}
+          labelType={labelType}
           placeholder={placeholder}
-          disabled={disabled}
-          precision={
-            typeof precision === 'number' && !isNaN(precision)
-              ? Math.round(precision)
-              : undefined
-          }
+          precision={precision}
           min={min}
           max={max}
-          className={{
-            ...className?.numberField,
-            input: twMerge(
-              meta.error && meta.touched && 'border-red-400 bg-red-50',
-              className?.numberField?.input
-            ),
+          tooltip={tooltip}
+          required={required}
+          hideError={hideError}
+          error={meta.error}
+          isTouched={meta.touched}
+          disabled={disabled}
+          onBlur={() => {
+            helpers.setTouched(true)
+            onBlur?.()
           }}
+          data={data}
+          className={className}
+          {...props}
         />
       </div>
-      {!hideError && meta.touched && meta.error && (
-        <div
-          className={twMerge(
-            'w-full text-right text-sm text-red-400',
-            className?.error
-          )}
-        >
-          {meta.error}
-        </div>
-      )}
+    )
+  }
+
+  return (
+    <div className={twMerge('w-full', className?.root)}>
+      <NumberField
+        id={id}
+        value={value!}
+        onChange={onChange!}
+        label={label}
+        labelType={labelType}
+        placeholder={placeholder}
+        precision={precision}
+        min={min}
+        max={max}
+        tooltip={tooltip}
+        required={required}
+        hideError={hideError}
+        error={error}
+        isTouched={isTouched}
+        disabled={disabled}
+        onBlur={() => {
+          helpers.setTouched(true)
+          onBlur?.()
+        }}
+        data={data}
+        className={className}
+        {...props}
+      />
     </div>
   )
 }
