@@ -1,6 +1,10 @@
-import { faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons'
+import {
+  faEye,
+  faEyeSlash,
+  faMagnifyingGlass,
+} from '@fortawesome/free-solid-svg-icons'
 import { Form, Formik } from 'formik'
-import React, { useState } from 'react'
+import { useState } from 'react'
 import * as yup from 'yup'
 import Button from '../Button'
 import FormikTextField from './FormikTextField'
@@ -26,7 +30,7 @@ export const Default = () => (
                 name="name"
                 label="Label"
                 tooltip="Tooltip for this input"
-                className={{ root: 'mb-1' }}
+                className={{ root: 'mb-1 w-80' }}
                 placeholder="Placeholder"
               />
               <Button type="submit">Submit</Button>
@@ -38,6 +42,86 @@ export const Default = () => (
     </Formik>
   </div>
 )
+
+export const Icon = () => (
+  <div>
+    <Formik
+      initialValues={{
+        name: '',
+      }}
+      isInitialValid={false}
+      onSubmit={async (values, { resetForm }) => {
+        alert(`Form submitted with value: ${values.name}`)
+        resetForm()
+      }}
+    >
+      {({ values }) => {
+        return (
+          <div>
+            <Form>
+              <FormikTextField
+                name="name"
+                label="Label"
+                tooltip="Tooltip for this input"
+                className={{ root: 'mb-1 w-80' }}
+                placeholder="Placeholder"
+                icon={faMagnifyingGlass}
+                onIconClick={() => alert('Icon clicked')}
+              />
+              <Button type="submit">Submit</Button>
+            </Form>
+            <div>Value: {values.name}</div>
+          </div>
+        )
+      }}
+    </Formik>
+  </div>
+)
+
+export const IconRight = () => {
+  const [textHidden, setTextHidden] = useState(true)
+
+  return (
+    <div>
+      <div>
+        The icon prop allows to pass a fontawesom icon to the component, which
+        will then be displayed on the right side of the input field
+      </div>
+      <Formik
+        initialValues={{
+          name: '',
+        }}
+        isInitialValid={false}
+        onSubmit={async (values, { resetForm }) => {
+          alert(`Form submitted with value: ${values.name}`)
+          resetForm()
+        }}
+      >
+        {({ values }) => {
+          return (
+            <div>
+              <Form>
+                <FormikTextField
+                  name="name"
+                  label="Label"
+                  tooltip="Tooltip for this input"
+                  className={{ root: 'mb-1' }}
+                  placeholder="Placeholder"
+                  icon={textHidden ? faEye : faEyeSlash}
+                  iconPosition="right"
+                  onIconClick={() => setTextHidden(!textHidden)}
+                  type={textHidden ? 'password' : 'text'}
+                />
+                <Button type="submit">Submit</Button>
+              </Form>
+              <div>Value: {values.name}</div>
+            </div>
+          )
+        }}
+      </Formik>
+    </div>
+  )
+}
 
 export const Disabled = () => (
   <div>
@@ -171,11 +255,11 @@ export const OnChangeError = () => (
           <div>
             <Form>
               <FormikTextField
+                isTouched
                 error="Error message"
-                hideError={false}
                 value={values.name}
                 onChange={(newValue) => {
-                  setFieldValue('name', newValue.replace(/\s/g, ''))
+                  setFieldValue('name', newValue)
                 }}
                 label="Label"
                 tooltip="Tooltip for this input"
@@ -274,12 +358,9 @@ export const Validation = () => (
   </div>
 )
 
-export const SmallLabel = () => (
+export const LargeLabel = () => (
   <div>
-    <div>
-      Formik text area component with a small label (designed e.g. for login
-      forms)
-    </div>
+    <div>Formik text area component with a large label</div>
     <Formik
       initialValues={{
         name: '',
@@ -298,7 +379,7 @@ export const SmallLabel = () => (
                 required
                 name="name"
                 label="Label"
-                labelType="small"
+                labelType="large"
                 tooltip="Tooltip for this input"
                 className={{ root: 'mb-1' }}
                 placeholder="Placeholder"
@@ -312,47 +393,3 @@ export const SmallLabel = () => (
     </Formik>
   </div>
 )
-
-export const Icon = () => {
-  const [textHidden, setTextHidden] = useState(true)
-
-  return (
-    <div>
-      <div>
-        The icon prop allows to pass a fontawesom icon to the component, which
-        will then be displayed on the right side of the input field
-      </div>
-      <Formik
-        initialValues={{
-          name: '',
-        }}
-        isInitialValid={false}
-        onSubmit={async (values, { resetForm }) => {
-          alert(`Form submitted with value: ${values.name}`)
-          resetForm()
-        }}
-      >
-        {({ values }) => {
-          return (
-            <div>
-              <Form>
-                <FormikTextField
-                  name="name"
-                  label="Label"
-                  tooltip="Tooltip for this input"
-                  className={{ root: 'mb-1' }}
-                  placeholder="Placeholder"
-                  icon={textHidden ? faEye : faEyeSlash}
-                  onIconClick={() => setTextHidden(!textHidden)}
-                  type={textHidden ? 'password' : 'text'}
-                />
-                <Button type="submit">Submit</Button>
-              </Form>
-              <div>Value: {values.name}</div>
-            </div>
-          )
-        }}
-      </Formik>
-    </div>
-  )
-}
