@@ -18,7 +18,6 @@ export interface LabelProps {
   showTooltipSymbol?: boolean
   tooltipSymbolSize?: 'sm' | 'md' | 'lg' | 'xl'
   className?: {
-    override?: string
     root?: string
     tooltip?: string
     tooltipSymbol?: string
@@ -67,11 +66,7 @@ export function Label({
             data-cy={data?.cy}
             data-test={data?.test}
             htmlFor={forId}
-            className={twMerge(
-              className?.override,
-              'cursor-default',
-              className?.root
-            )}
+            className={twMerge('cursor-default', className?.root)}
           >
             {label}
           </RadixLabel.Root>
@@ -89,11 +84,7 @@ export function Label({
           data-cy={data?.cy}
           data-test={data?.test}
           htmlFor={forId}
-          className={twMerge(
-            className?.override,
-            'mr-2 cursor-default',
-            required && 'mr-0'
-          )}
+          className={twMerge('mr-2 cursor-default', required && 'mr-0')}
         >
           {label}
         </RadixLabel.Root>
@@ -128,7 +119,6 @@ export function Label({
         data-test={data?.test}
         htmlFor={forId}
         className={twMerge(
-          className?.override,
           'flex cursor-default flex-row items-center',
           className?.root
         )}
