@@ -1,33 +1,40 @@
+import { faCircleExclamation } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import React from 'react'
 import { twMerge } from 'tailwind-merge'
-import Label from './Label'
+import FormLabel from '../FormLabel'
+import { Tooltip } from '../Tooltip'
 
 export interface NumberFieldClassName {
-  override?: string
-  root?: string
-  input?: string
+  field?: string
   label?: string
+  input?: string
+  error?: string
   tooltip?: string
 }
 
 export interface NumberFieldProps {
   id?: string
+  value: string | number
+  onChange: (newValue: string) => void
+  label?: string
+  labelType?: 'small' | 'large'
+  placeholder?: string
+  precision?: number
+  min?: number
+  max?: number
+  tooltip?: string | React.ReactNode
+  required?: boolean
+  hideError?: boolean
+  error?: string
+  isTouched?: boolean
+  disabled?: boolean
+  onBlur?: () => void
   data?: {
     cy?: string
     test?: string
   }
-  value: string | number
-  onChange: (newValue: string) => void
-  label?: string
-  tooltip?: string | React.ReactNode
-  required?: boolean
-  onBlur?: () => void
-  placeholder?: string
-  disabled?: boolean
   className?: NumberFieldClassName
-  precision?: number
-  min?: number
-  max?: number
   [key: string]: any
 }
 
@@ -35,38 +42,46 @@ export interface NumberFieldProps {
  * This function returns a text field component for use without formik
  *
  * @param id - The id of the input field.
- * @param data - The object of data attributes that can be used for testing (e.g. data-test or data-cy)
  * @param value - The value of the input field (external state management).
  * @param onChange - The onChange function of the input field (external state management).
  * @param label - The text displayed as label.
- * @param tooltip - The optional tooltip is shown on hover over the tooltip next to the label.
- * @param required - Indicate whether the field is required or not.
- * @param onBlur - The onBlur function of the input field.
+ * @param labelType - The optional labelType can be used to change the size and position of the label according to pre-defined standards.
  * @param placeholder - The placeholder text for the input field.
- * @param disabled - Indicate whether the field is disabled or not.
  * @param precision - The optional precision defines the number of decimal places that are allowed.
  * @param min - The optional min defines the minimum value that is allowed.
  * @param max - The optional max defines the maximum value that is allowed.
+ * @param tooltip - The optional tooltip is shown on hover over the tooltip next to the label.
+ * @param required - Indicate whether the field is required or not.
+ * @param hideError - Indicate whether the error message should be hidden or not.
+ * @param error - The error message that is displayed below the input field.
+ * @param isTouched - Indicate whether the field has been touched or not (validation is not handled by this component).
+ * @param disabled - Indicate whether the field is disabled or not.
+ * @param onBlur - The onBlur function of the input field.
+ * @param data - The object of data attributes that can be used for testing (e.g. data-test or data-cy)
  * @param className - The optional className object allows you to override the default styling.
  */
-
 export function NumberField({
   id,
-  data,
   value,
   onChange,
   label,
-  tooltip,
-  required,
-  onBlur,
+  labelType = 'small',
   placeholder,
-  disabled,
   precision,
   min,
   max,
+  tooltip,
+  required = false,
+  hideError,
+  error,
+  isTouched,
+  disabled,
+  onBlur,
+  data,
   className,
+  ...props
 }: NumberFieldProps): React.ReactElement {
-  const regex =
+  const validInput =
     typeof precision === 'number' && !isNaN(precision)
       ? precision === 0
         ? /^[-]?\d*$/
@@ -74,52 +89,72 @@ export function NumberField({
       : /^[-]?\d*\.?\d*$/
 
   return (
-    <div className={twMerge('flex flex-row', className?.root)}>
+    <div
+      className={twMerge(
+        'flex w-full flex-row',
+        labelType === 'small' && 'flex-col',
+        className?.field
+      )}
+    >
       {label && (
-        <Label
-          forId={id}
+        <FormLabel
+          id={id}
           required={required}
           label={label}
-          className={{
-            root: twMerge('my-auto mr-2 min-w-max font-bold', className?.label),
-            tooltip: twMerge('text-sm font-normal', className?.tooltip),
-          }}
+          labelType={labelType}
           tooltip={tooltip}
-          showTooltipSymbol={typeof tooltip !== 'undefined'}
+          className={className}
         />
       )}
 
-      <input
-        id={id}
-        data-cy={data?.cy}
-        data-test={data?.test}
-        type="text"
-        value={value}
-        onChange={(e) => {
-          if (
-            e.target.value.match(regex) !== null &&
-            (e.target.value === '' ||
-              typeof min === 'undefined' ||
-              parseFloat(e.target.value) >= min) &&
-            (e.target.value === '' ||
-              typeof max === 'undefined' ||
-              parseFloat(e.target.value) <= max)
-          ) {
-            onChange(e.target.value)
-          } else {
-            console.log(`input ${e.target.value} does not match regex ${regex}`)
-          }
-        }}
-        onBlur={onBlur}
-        placeholder={placeholder}
-        disabled={disabled}
-        className={twMerge(
-          className?.override,
-          'focus:border-uzh-blue-50 h-9 w-full rounded border border-uzh-grey-60 bg-uzh-grey-20',
-          disabled && 'cursor-not-allowed',
-          className?.input
+      <div className="flex w-full flex-row items-center gap-2">
+        <input
+          id={id}
+          data-cy={data?.cy}
+          data-test={data?.test}
+          type="text"
+          value={value}
+          onChange={(e) => {
+            if (
+              e.target.value.match(validInput) !== null &&
+              (e.target.value === '' ||
+                typeof min === 'undefined' ||
+                parseFloat(e.target.value) >= min) &&
+              (e.target.value === '' ||
+                typeof max === 'undefined' ||
+                parseFloat(e.target.value) <= max)
+            ) {
+              onChange(e.target.value)
+            } else {
+              console.log(
+                `input ${e.target.value} does not match regex ${validInput}`
+              )
+            }
+          }}
+          onBlur={onBlur}
+          placeholder={placeholder}
+          disabled={disabled}
+          className={twMerge(
+            'focus:border-uzh-blue-50 h-9 w-full rounded border border-uzh-grey-60 pl-2 placeholder-slate-400',
+            disabled && 'cursor-not-allowed',
+            !!error && isTouched && 'border-red-400 bg-red-50',
+            className?.input
+          )}
+          {...props}
+        />
+        {error && !hideError && isTouched && (
+          <Tooltip
+            tooltip={error}
+            delay={0}
+            className={{ tooltip: 'max-w-[30rem] text-sm' }}
+          >
+            <FontAwesomeIcon
+              icon={faCircleExclamation}
+              className="mr-1 text-red-600"
+            />
+          </Tooltip>
         )}
-      />
+      </div>
     </div>
   )
 }
